requirements.txt
python-telegram-bot
google-generativeai
python-dotenv
aiohttp
Pillow
pytest
pytest-asyncio
